tweak to IEC serial protocol change Debug_print* macros

+++ lib/meatloaf/tape/t64.cpp
@@ -126,10 +126,10 @@ bool T64MStream::seekPath(std::string path) {
     {
         //auto entry = containerImage->entry;
         auto type = decodeType(entry.file_type).c_str();
-        size_t start_address = UINT16_FROM_HILOBYTES(entry.start_address[1], entry.start_address[0]);
-        size_t end_address = UINT16_FROM_HILOBYTES(entry.end_address[1], entry.end_address[0]);
-        size_t data_offset = UINT32_FROM_LE_UINT32(entry.data_offset);
-        Debug_printv("filename [%.16s] type[%s] start_address[%zu] end_address[%zu] data_offset[%zu]", entry.filename, type, start_address, end_address, data_offset);
+        uint32_t start_address = UINT16_FROM_HILOBYTES(entry.start_address[1], entry.start_address[0]);
+        uint32_t end_address = UINT16_FROM_HILOBYTES(entry.end_address[1], entry.end_address[0]);
+        uint32_t data_offset = UINT32_FROM_LE_UINT32(entry.data_offset);
+        Debug_printv("filename [%.16s] type[%s] start_address[%lu] end_address[%lu] data_offset[%lu]", entry.filename, type, start_address, end_address, data_offset);
 
         // Calculate file size
         _size = ( end_address - start_address ) + 2; // 2 bytes for load address
@@ -142,7 +142,7 @@ bool T64MStream::seekPath(std::string path) {
         _position = 0;
         containerStream->seek(entry.data_offset);
 
-        Debug_printv("File Size: size[%d] available[%d] position[%d]", _size, available(), _position);
+        Debug_printv("File Size: size[%ld] available[%ld] position[%ld]", _size, available(), _position);
 
         return true;
     }
